create ActionMenu component
