+++ frontend/src/metabase/collections/components/ActionMenu/ActionMenu.styled.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+import EntityItem from "metabase/components/EntityItem";
+import { color } from "metabase/lib/colors";
+
+export const EntityItemMenu = styled(EntityItem.Menu)`
+  color: ${color("text-medium")};
+`;
