+++ frontend/src/metabase/collections/components/ActionMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ActionMenu";
